// src/models/tootmine/tootmineModels.ts
export interface gruppTootajad {
  PNIMI: string;
  Y: string;
  JRK: number;
  GGRUPP: string;
  GNIMI: string;
  pilt: string;
  RID: number;
  JID: number;
  AegKokku: number;
  GID: number;
  TOO: string;
  KOGUS: string;
  LEPNR: string;
  Nimi: string;
  START: string;
  ENIMI: string;
  TID: number;
  Markus: string;
}
export interface toolGrupp {
  GGRUPP: string;
  JRK: number;
  Kokku: number;
  asukoht_id: number;
}
export interface Tulem {
  tulem: number;
}